separate cli subcommands into different namespaces
